+++ crates/lang/tests/ui/contract/pass/constructor-payable.rs
@@ -1,0 +1,23 @@
+use ink_lang as ink;
+
+#[ink::contract]
+mod contract {
+    #[ink(storage)]
+    pub struct Contract {}
+
+    impl Contract {
+        #[ink(constructor, selector = 0, payable)]
+        pub fn constructor() -> Self {
+            Self {}
+        }
+
+        #[ink(message)]
+        pub fn message(&self) {}
+    }
+}
+
+use contract::Contract;
+
+fn main() {
+    assert!(<Contract as ::ink_lang::reflect::DispatchableConstructorInfo<0>>::PAYABLE);
+}