+++ RELEASES.md
@@ -4,6 +4,7 @@ This is the 8th release candidate for ink! 3.0.
 
 ## Change
 - Renamed the `ink_env` function `transferred_balance()` to `transferred_value()` ‒ [#1063](https://github.com/paritytech/ink/pull/1063).
+- Allow specifying payable constructors - [#1065](https://github.com/paritytech/ink/pull/1065)
 
 # Version 3.0-rc7